Add a copy link to github branches in the attribute to make it easier to switch to that branch
