+++ src/components/ExternalLink.js
@@ -12,10 +12,10 @@ const isValidExternalLink = (urlString) => {
 /**
  * Ensures that Github target _blank links have the required noopener noreferrer properties
  * and that they point to Github. Use like a normal <a> tag
- * 
+ *
  * @type {React.FC<{href: string}>}
  */
-const ExternalLink = (props) => {
+export const ExternalLink = (props) => {
   if (isValidExternalLink(props.href)) {
     return (
       <a {...props} target="_blank" rel="noopener noreferrer">
@@ -24,5 +24,3 @@ const ExternalLink = (props) => {
     );
   } else return <>"Invalid external URL."</>;
 };
-
-export default ExternalLink;
