+++ src/lib/fields.js
@@ -3,6 +3,10 @@ const PULL_REQUESTS_FIELD = "pullRequests";
 const BRANCHES_FIELD = "branches";
 
 /**
+ * @typedef {Aha.ReferenceInterface & Aha.HasExtensionFields} LinkableRecord
+ */
+
+/**
  * @typedef PrLink
  * @prop {number} id
  * @prop {string} name
@@ -18,16 +22,17 @@ const BRANCHES_FIELD = "branches";
  */
 
 /**
- * Append a field/value pair to the given record. Returns an actual record
- * instance if one existed.
+ * Append a field/value pair to the given record.
  *
- * @param {Aha.ApplicationModel} record
+ * @param {Aha.ApplicationModel & Aha.HasExtensionFields} record
  * @param {string} fieldName
  * @param {*} newValue
  */
 async function appendField(record, fieldName, newValue) {
   // Link to Aha! record.
-  console.log(`Link to ${record.typename}:${record.referenceNum || record.id}`);
+  console.log(
+    `Link to ${record.typename}:${record["referenceNum"] || record.uniqueId}`
+  );
 
   await replaceField(record, fieldName, (value) => {
     /** @type {{id:any}[]} */
@@ -46,7 +51,7 @@ async function appendField(record, fieldName, newValue) {
 
 /**
  * @template T
- * @param {Aha.ApplicationModel} record
+ * @param {Aha.HasExtensionFields} record
  * @param {string} fieldName
  * @param {((value: T|null) => T | Promise<T>)} replacer
  */
@@ -79,7 +84,7 @@ function githubPrToPrLink(pr) {
 
 /**
  * @param {import("./github").PrForLink} pr
- * @param {Aha.RecordStub} record
+ * @param {LinkableRecord} record
  */
 async function linkPullRequestToRecord(pr, record) {
   await appendField(record, PULL_REQUESTS_FIELD, githubPrToPrLink(pr));
@@ -89,6 +94,10 @@ async function linkPullRequestToRecord(pr, record) {
     prNumber: pr.number,
     ahaReference: [record.typename, record.referenceNum],
   });
+
+  if (pr.headRef) {
+    await linkBranchToRecord(pr.headRef.name, pr.repository.url, record);
+  }
 }
 
 /**
@@ -105,7 +114,7 @@ async function linkPullRequest(pr) {
 }
 
 /**
- * @param {Aha.RecordStub} record
+ * @param {LinkableRecord} record
  * @param {*} number
  */
 async function unlinkPullRequest(record, number) {
@@ -129,7 +138,7 @@ async function unlinkPullRequest(record, number) {
 }
 
 /**
- * @param {Aha.RecordStub} record
+ * @param {Aha.ReferenceInterface & Aha.HasExtensionFields} record
  */
 async function unlinkPullRequests(record) {
   /** @type {PrLink[]} */
@@ -137,12 +146,14 @@ async function unlinkPullRequests(record) {
     (await record.getExtensionField(IDENTIFIER, PULL_REQUESTS_FIELD)) || [];
   const ids = prs.map((pr) => accountPrId(pr.id, record.referenceNum));
 
-  await replaceField(aha.account, PULL_REQUESTS_FIELD, (
-    /** @type {AccountPr[]} */ accountPrs
-  ) => {
-    if (!accountPrs) return [];
-    return accountPrs.filter((accountPr) => !ids.includes(accountPr.id));
-  });
+  await replaceField(
+    aha.account,
+    PULL_REQUESTS_FIELD,
+    (/** @type {AccountPr[]} */ accountPrs) => {
+      if (!accountPrs) return [];
+      return accountPrs.filter((accountPr) => !ids.includes(accountPr.id));
+    }
+  );
 
   await record.setExtensionField(IDENTIFIER, PULL_REQUESTS_FIELD, []);
 }
@@ -156,22 +167,31 @@ export async function allPrs() {
 }
 
 /**
+ * @param {LinkableRecord} record
+ * @param {string} branchName
+ * @param {string} repoUrl
+ */
+async function linkBranchToRecord(branchName, repoUrl, record) {
+  await appendField(record, BRANCHES_FIELD, {
+    id: branchName,
+    name: branchName,
+    url: `${repoUrl}/tree/${branchName}`,
+  });
+}
+
+/**
  * @param {string} branchName
  * @param {string} repoUrl
  */
 async function linkBranch(branchName, repoUrl) {
   const record = await referenceToRecord(branchName);
   if (record) {
-    await appendField(record, BRANCHES_FIELD, {
-      id: branchName,
-      name: branchName,
-      url: `${repoUrl}/tree/${branchName}`,
-    });
+    return await linkBranchToRecord(branchName, repoUrl, record);
   }
 }
 
 /**
- * @param {Aha.RecordStub} record
+ * @param {Aha.HasExtensionFields} record
  */
 async function unlinkBranches(record) {
   await record.setExtensionField(IDENTIFIER, BRANCHES_FIELD, []);
@@ -179,7 +199,7 @@ async function unlinkBranches(record) {
 
 /**
  * @param {string} str
- * @returns {Promise<Aha.RecordStub|null>}
+ * @returns {Promise<Aha.ReferenceInterface & Aha.HasExtensionFields|null>}
  */
 async function referenceToRecord(str) {
   const ahaReference = extractReference(str);
