+++ src/links.js
@@ -1,38 +1,7 @@
-import { AuthProvider, useAuth } from "@aha-app/aha-develop-react";
+import { AuthProvider } from "@aha-app/aha-develop-react";
 import React from "react";
-import { render, unmountComponentAtNode } from "react-dom";
-import Branches from "./components/Branches";
-import Menu from "./components/Menu";
-import PullRequests from "./components/PullRequest";
 import Styles from "./components/Styles";
-
-function App({ fields, record }) {
-  const { error, authed } = useAuth(async () => {});
-  const authError = error && <div>{error}</div>;
-
-  const isLinked = [fields.branches, fields.pullRequests].some(
-    (ary) => ary?.length > 0
-  );
-
-  const githubLinks = isLinked ? (
-    <aha-flex direction="column" gap="8px" justify-content="space-between">
-      <Branches fields={fields} />
-      <PullRequests record={record} prs={fields.pullRequests}></PullRequests>
-    </aha-flex>
-  ) : (
-    <div>Not linked</div>
-  );
-
-  return (
-    <>
-      <aha-flex align-items="center" justify-content="space-between" gap="5px">
-        {authError}
-        {githubLinks}
-        <Menu record={record} />
-      </aha-flex>
-    </>
-  );
-}
+import { Attribute } from "./components/attribute/Attribute";
 
 /**
  * @type {Aha.RenderExtension}
@@ -42,7 +11,7 @@ const links = ({ record, fields }) => {
     <>
       <Styles />
       <AuthProvider serviceName="github" serviceParameters={{ scope: "repo" }}>
-        <App fields={fields} record={record} />
+        <Attribute fields={fields} record={record} />
       </AuthProvider>
     </>
   );
